requirements.txt
pytest==7.3.1
requests==2.31.0
PyYAML==6.0.1
allure-pytest==2.13.2